(PC-36379)[PRO] feat: improve domain cases in IndividualOfferPriceTable
