+++ pro/src/pages/IndividualOffer/IndividualOfferPriceTable/commons/constants.ts
@@ -1,3 +1,6 @@
+export const FAILED_PATCH_OFFER_USER_MESSAGE =
+  'Une erreur est survenue lors de la mise à jour de votre offre'
+
 export const PRICE_TABLE_ENTRY_MAX_LABEL_LENGTH = 50
 export const PRICE_TABLE_ENTRY_MAX_PRICE_IN_EUR = 300
 export const PRICE_TABLE_ENTRY_MAX_PRICE_IN_XPF = 20_000
