+++ pro/src/pages/IndividualOffer/IndividualOfferPriceTable/commons/utils/saveEventOfferPriceTable.ts
@@ -1,7 +1,8 @@
 import { api } from '@/apiClient/api'
 import type { GetIndividualOfferWithAddressResponseModel } from '@/apiClient/v1'
-import { FrontendError } from '@/commons/errors/FrontendError'
+import { handleError } from '@/commons/errors/handleError'
 
+import { FAILED_PATCH_OFFER_USER_MESSAGE } from '../constants'
 import type { PriceTableFormValues } from '../schemas'
 import { toPriceCategoryBody } from './toPriceCategoryBody'
 
@@ -15,17 +16,9 @@ export const saveEventOfferPriceTable = async (
 ) => {
   try {
     await api.patchOffer(offer.id, { isDuo: formValues.isDuo })
-  } catch {
-    throw new FrontendError(
-      'Une erreur est survenue lors de la mise à jour de votre tarif'
-    )
-  }
 
-  try {
     await api.postPriceCategories(offer.id, toPriceCategoryBody(formValues))
-  } catch {
-    throw new FrontendError(
-      'Une erreur est survenue lors de la mise à jour de votre tarif'
-    )
+  } catch (error) {
+    return handleError(error, FAILED_PATCH_OFFER_USER_MESSAGE)
   }
 }
