--- pro/src/pages/IndividualOffer/IndividualOfferPriceTable/components/ActivationCodeFormDialog/ActivationCodeFormDialog.module.scss
@@ -1,72 +1,0 @@
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_size.scss" as size;
-
-.activation-codes-upload {
-  text-align: center;
-
-  @media (min-width: size.$tablet) {
-    width: rem.torem(540px);
-  }
-
-  .activation-codes-errors {
-    margin-top: rem.torem(8px);
-    margin-bottom: rem.torem(16px);
-
-    p {
-      margin-bottom: rem.torem(8px);
-    }
-  }
-
-  .activation-codes-upload-description {
-    margin-top: rem.torem(16px);
-    margin-bottom: rem.torem(16px);
-  }
-
-  .activation-codes-upload-button-caption {
-    @include fonts.body-accent-xs;
-
-    margin-top: rem.torem(16px);
-    color: var(--color-text-subtle);
-  }
-
-  .activation-codes-upload-separator {
-    border: rem.torem(1px) solid var(--color-border-subtle);
-    margin-bottom: rem.torem(16px);
-    margin-top: rem.torem(24px);
-    width: 100%;
-  }
-
-  .activation-codes-upload-template-section {
-    text-align: left;
-    width: 100%;
-  }
-
-  .activation-codes-upload-gabarit {
-    @include fonts.title4;
-
-    margin-bottom: rem.torem(16px);
-  }
-
-  .activation-codes-form {
-    p {
-      margin-bottom: rem.torem(8px);
-    }
-  }
-
-  .activation-codes-form-expiration-date {
-    margin: rem.torem(16px) auto;
-    max-width: rem.torem(250px);
-  }
-
-  .activation-codes-actions-button {
-    display: flex;
-    justify-content: space-between;
-    text-align: center;
-    margin-top: rem.torem(16px);
-  }
-
-  .date-input svg {
-    width: rem.torem(20px);
-  }
-}
