+++ pro/src/pages/IndividualOffer/IndividualOfferPriceTable/components/IndividualOfferPriceTableScreen.module.scss
@@ -1,58 +1,5 @@
 @use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_size.scss" as size;
 
-.row {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  margin-bottom: rem.torem(16px);
-}
-
-.callout-area {
-  margin-bottom: 8px;
-}
-
-.callout-area-margin {
-  margin-top: rem.torem(32px);
-}
-
-.button-actions {
-  display: flex;
-  gap: rem.torem(16px);
-  flex-direction: column;
-  margin-bottom: rem.torem(16px);
-}
-
-.button-action {
-  justify-content: start;
-}
-
-@media (min-width: size.$tablet) {
-  .row {
-    flex-direction: row;
-    gap: rem.torem(16px);
-    align-items: flex-start;
-  }
-
-  .field-layout-xsmall {
-    max-width: rem.torem(108px);
-  }
-
-  .field-layout-small {
-    max-width: rem.torem(224px);
-  }
-
-  .field-layout-shrink {
-    width: initial;
-  }
-
-  .field-layout-footer {
-    min-height: rem.torem(32px);
-  }
-
-  .button-actions {
-    flex-direction: row;
-    margin-top: rem.torem(40px);
-  }
+.callout {
+  margin-bottom: rem.torem(24px);
 }
